increasing number of props and enemies

+++ game/Core/States/Arena.ts
@@ -51,6 +51,12 @@ module SpaceWars.Core.States {
         enemies: SpaceWars.Collections.EnemyGroup;
         // Time at which next enemy will enter the arena
         nextEnemySpawn: number;
+
+				///////////////////////////////////
+				/* ##### Performance Enha. ##### */
+				///////////////////////////////////
+
+				slowedLoopNext: number = 0;
         
         ///////////////////////////////////
         /* ##### The UserInterface ##### */
@@ -109,7 +115,7 @@ module SpaceWars.Core.States {
 			this.combo = combo;
 
             // Settings for the Game should be set here
-            this.totalEnemies = 20;
+            this.totalEnemies = 70;
 
         }
 
@@ -141,7 +147,8 @@ module SpaceWars.Core.States {
             this.generateEnemies();
 
             // Spawning Enemies with timed interval
-            this.nextEnemySpawn = this.game.time.now + 2000;
+            this.nextEnemySpawn = this.game.time.now + 1000;
+            this.game.time.events.repeat(Phaser.Timer.SECOND * 1, this.totalEnemies, this.spawnEnemy, this);
             this.game.time.events.repeat(Phaser.Timer.SECOND * 2, this.totalEnemies, this.spawnEnemy, this);
 
             // Loading UI
@@ -158,6 +165,20 @@ module SpaceWars.Core.States {
 
 
         }
+
+				ai() {
+				
+             //#### AI #### //
+
+            // try to get player in sight
+            this.seekPlayerY();
+
+            // When player is in sight of enemy
+            var enemy = this.playerInSight();
+            if(enemy) {
+                enemy.shoot();
+            }
+				}
 
         update()
         {
@@ -179,41 +200,38 @@ module SpaceWars.Core.States {
                 this.game.physics.arcade.overlap(this.bullets, this.player, this.playerHit, null, this);
             }, this);
 
+						if(this.slowedLoopNext <= this.game.time.now) {
+							this.slowedLoopNext = this.game.time.now + 200; 
+							this.ai();
 
-             //#### AI #### //
+							// Killing Props and enemies when they hit left world boundry
+							this.terrainProps.forEachAlive(function (prop)
+							{
+									if (prop.body.x + prop.body.width < 0) {
+											prop.kill();
+									} 
+							}, this);
 
-            // try to get player in sight
-            this.seekPlayerY();
+							//// TODO: Limit foreach to enemies on screen
+							this.enemies.forEachExists(function (enemy)
+							{
+									if (enemy.x + enemy.width < 0) {
+											// Resetting position
+											enemy.body.x = this.game.world.width + enemy.body.width;
+											enemy.body.y = this.game.rnd.integerInRange(this.game.world.y, this.game.world.height);
+											enemy.body.velocity.x = -200;
+									}
+							}, this);
+						}
 
-            // When player is in sight of enemy
-            var enemy = this.playerInSight();
-            if(enemy) {
-                enemy.shoot();
-            }
+
 
             // ### SPAWN TERRAIN PROPS ### //
             if (this.game.time.now > this.nextPropSpawn) {
                 this.spawnProp();
             }
 
-            // Killing Props and enemies when they hit left world boundry
-            this.terrainProps.forEachAlive(function (prop)
-            {
-                if (prop.body.x + prop.body.width < 0) {
-                    prop.kill();
-                } 
-            }, this);
             
-            //// TODO: Limit foreach to enemies on screen
-            this.enemies.forEachExists(function (enemy)
-            {
-                if (enemy.x + enemy.width < 0) {
-                    // Resetting position
-                    enemy.body.x = this.game.world.width + enemy.body.width;
-                    enemy.body.y = this.game.rnd.integerInRange(this.game.world.y, this.game.world.height);
-                    enemy.body.velocity.x = -200;
-                }
-            }, this);
 
 
 
@@ -274,7 +292,7 @@ module SpaceWars.Core.States {
 
             }
 
-            this.nextPropSpawn = this.game.time.now + 1000;
+            this.nextPropSpawn = this.game.time.now + 300;
         }
 
 
@@ -284,14 +302,20 @@ module SpaceWars.Core.States {
             // Grabbing a prop
             var prop = this.terrainProps.getFirstDead();
 
+						prop.anchor.y = 1.0;
+						prop.anchor.x = 1.0;
+
+						var rand = this.game.rnd.realInRange(0, 3);
+						prop.scale.setTo(rand, rand);
+
             // Resetting position
-            prop.reset(this.game.world.width + prop.body.width, this.game.world.height - prop.body.height);
+            prop.reset(this.game.world.width + prop.body.width, this.game.world.height);
 
             // Setting Movement Speed on prop 
             prop.body.velocity.x = -200;
 
             // When is next spawn
-            this.nextPropSpawn = this.game.time.now + this.game.rnd.integerInRange(500, 2000); 
+            this.nextPropSpawn = this.game.time.now + this.game.rnd.integerInRange(100, 300); 
         }
 
         // #### User Interface ####
@@ -480,12 +504,14 @@ module SpaceWars.Core.States {
         {
             // Adding explosion to game
             var explosion = this.game.add.sprite(sprite.body.x - (sprite.body.width / 2), sprite.body.y, 'explosion_1');
+            this.game.physics.enable(explosion, Phaser.Physics.ARCADE);
             explosion.anchor.x = 0.5;
             explosion.anchor.y = 0.5;
 
             explosion.animations.add('explode');
 
-            explosion.animations.play('explode', 80, false, true);
+            explosion.animations.play('explode', 10, false, true);
+            explosion.body.velocity.x = -255;
 
         }
